requirements.txt

streamlit>=1.39.0
llama_cpp_python==0.3.1
python-dotenv==1.0.1
# psycopg2==2.9.10  # Standard version; requires PostgreSQL development libraries
psycopg2-binary==2.9.10  # Precompiled version for easier installation
